pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["BonesetterB <122158929+BonesetterB@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.1"
uvicorn = {extras = ["standard"], version = "^0.23.1"}
sqlalchemy = "^2.0.19"
alembic = "^1.11.1"
pydantic = "^2.1.1"
asyncpg = "^0.28.0"
fastapi-mail = "^1.4.1"
redis = "4.5.0"
cloudinary = "^1.34.0"
aiosqlite = "^0.19.0"
fastapi-limiter = "^0.1.5"
python-jose = "^3.3.0"
passlib = "^1.7.4"
libgravatar = "^1.0.4"
python-multipart = "^0.0.6"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.2"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
httpx = "^0.24.1"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath="."